Теперь данные берутся с сервера, удалил все ранее используемые данные и необходимые к ним функции
